+++ R/summarise_plus.R
@@ -155,7 +155,7 @@ summarise_plus <- function(data_frame,
                            values,
                            statistics = c("sum", "freq"),
                            formats    = list(),
-                           types      = c(),
+                           types      = "",
                            weight     = NULL,
                            nesting    = "deepest",
                            merge_back = FALSE,
@@ -313,6 +313,45 @@ summarise_plus <- function(data_frame,
         message(" X ERROR: No <values> variables provided. Summarise will be aborted.")
         return(invisible(NULL))
     }
+
+    #++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+    # Types
+    #++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+
+    # Get row variables from provided combinations
+    type_vars <- unlist_variables(types)
+
+    # If there are type provided
+    if (is.null(type_vars)){
+        message(" ! WARNING: <Types> must be provided in quotation marks. Types will be ignored.")
+        types <- NULL
+    }
+    else if (length(type_vars) > 0){
+        # Exclude total because it will most likely not be part of the data frame
+        type_vars <- type_vars[type_vars != "total"]
+        type_vars <- data_frame |> part_of_df(type_vars, check_only = TRUE)
+
+        if (is.list(type_vars)){
+            message(" ! WARNING: The provided <type> '", paste(type_vars[[1]], collapse = ", "), "' is not part of\n",
+                    "            the data frame. Types will be ignored.")
+            types <- NULL
+        }
+
+        invalid_types <- type_vars[!type_vars %in% class]
+
+        if (length(invalid_types) > 0){
+            message(" ! WARNING: The provided <type> '", paste(type_vars[[1]], collapse = ", "), "' is not part of\n",
+                    "            the <class> variables. Types will be ignored.")
+            types <- NULL
+        }
+
+        rm(invalid_types)
+    }
+    else{
+        types <- NULL
+    }
+
+    rm(type_vars)
 
     #++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
     # Double entries